relocated player to new class
